fix(#14): Moved utils to src folder
