+++ package.json
@@ -21,6 +21,7 @@
     "next-i18next": "^14.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-hook-form": "^7.45.4",
     "react-i18next": "^13.0.2",
     "typescript": "5.1.6"
   },
