+++ src/app/[lang]/home/page.tsx
@@ -1,19 +1,24 @@
 import React from 'react';
-import { getTranslation } from '../../../utils/getTranslation';
+import { getTranslation } from '../../../../locales/utils/getTranslation';
 import styles from './styles.module.scss';
 import { Input } from '@/components/UI/Input';
 
 interface HomeProps {
   params: any
 }
-async function Home({params}: HomeProps) {
+async function Home({ params }: HomeProps) {
   const lang = await getTranslation(params.lang)
-  return  (
-        <Input 
+  const handleSubmit = () => {
+  }
+  return (
+    <form onSubmit={handleSubmit}>
+      <Input
         type='text'
         placeholder={lang.nameInput}
-        />
-    )
+        name={'nickname'}
+      />
+    </form>
+  )
 }
 
 export default Home;
