+++ src/components/UI/Input/index.tsx
@@ -1,11 +1,15 @@
 import React from "react";
+import { useFormContext } from 'react-hook-form';
 import styles from './styles.module.scss';
 import { InputHTMLAttributes } from "react";
 
-interface InputProps extends InputHTMLAttributes<HTMLInputElement>{}
+interface InputProps extends InputHTMLAttributes<HTMLInputElement>{
+    name: string;
+}
 
-export function Input({...rest}: InputProps){
+export function Input({name ,...rest}: InputProps){
+    const { register } = useFormContext();
     return(
-        <input className={styles.input} {...rest}/>
+        <input className={styles.input} id={name} {...register(name)} {...rest}/>
     )
 }
